Add ALPACAS options to UI
Adds options in UI-related files.

Side note: These files are seeming a little cluttered and annoying to
work with. It is probably a good idea to remove all of the the unused
portions of the demo code sooner than later. They are really just in
the way at this point.