Rework render.tscn
This adds a sript to the Render scene and a label that will be used
to display the name of the current Evennia room the player is in. The
label essentially required a font to be imported into the project.

Additionally, the initial plans for the artwork directory to be
completely ignored by git has been scrapped. Some artwork will be
included here in the art/default directory. All other files in the art
directory will still be excluded.