Add basic object rendering
This allows objects to be rendered and unrendered as instances of
interactive.tscn on the fly based on messages from the server when an
object moves.

Note that the current parent node of the created instances does not
properly handle multiple objects. Everything is overlapping at the
center until this node is specialized handle this properly.