Move Player node into Scenery container
To accomplish this, I needed to change the Player node in various
ways. This included changing it's texture from a Sprite object to a
TextureRext object.

Additionally, I figured out how to get the Player object to render at
the bottom left inside of the Container objects by making a dummy
Control object that expands inside another container. This may
accomplish what I was trying to figure out with the exits earlier.