Account for key_name dupes of rendered objects
This allows a user to click on an object rendered in a scene that
contains multiple objects with dupicate key names. While working
through this, I thought it might be best to avoid exposing the
internal Evennia object ID to the client for objects. Therefore, I
decided to use a client side ID for the object via Godot's
"get_instance_id" built-in function instead. This seems to have
worked, although it could possible lead to some inconsistent
results that could be avoided using the internal object ID.